Remove un-needed appsettings artifacts
Rearrange SecurityHelper items
Rename a few items and update some whitespace
