+++ src/G42.UmbracoGrease/G42TransformationHelper/TransformationHelper.cs
@@ -22,6 +22,7 @@ namespace G42.UmbracoGrease.G42TransformationHelper
                 var viewResult = System.Web.Mvc.ViewEngines.Engines.FindPartialView(context, viewName);
                 var viewContext = new ViewContext(context, viewResult.View, new ViewDataDictionary(model), new TempDataDictionary(), sw);
                 viewResult.View.Render(viewContext, sw);
+
                 return sw.GetStringBuilder().ToString();
             }
         }
