+++ src/G42.UmbracoGrease/Helpers/IpHelper.cs
@@ -11,12 +11,14 @@
         /// <returns></returns>
         public static string GetIpAddress()
         {
-            System.Web.HttpContext context = System.Web.HttpContext.Current;
-            string ipAddress = context.Request.ServerVariables["HTTP_X_FORWARDED_FOR"];
+            var context = System.Web.HttpContext.Current;
+            
+            var ipAddress = context.Request.ServerVariables["HTTP_X_FORWARDED_FOR"];
 
             if (!string.IsNullOrEmpty(ipAddress))
             {
-                string[] addresses = ipAddress.Split(',');
+                var addresses = ipAddress.Split(',');
+                
                 if (addresses.Length != 0)
                 {
                     return addresses[0];
