+++ src/G42.UmbracoGrease/Helpers/SecurityHelper.cs
@@ -1,5 +1,6 @@
 ﻿using System.Security.Cryptography;
 using System.Text;
+using System.Web;
 
 namespace G42.UmbracoGrease.Helpers
 {
@@ -27,6 +28,41 @@ namespace G42.UmbracoGrease.Helpers
             }
 
             return sb.ToString();
+        }
+
+        /// <summary>
+        /// Ensures the AWS SSL when using a shared SSL environment.
+        /// </summary>
+        /// <param name="enableLocal">if set to <c>true</c> [enable local].</param>
+        public static void EnsureAwsSsl(bool enableLocal = false)
+        {
+            if (HttpContext.Current.Request.Url.Host.EndsWith(".local") && !enableLocal)
+            {
+                return;
+            }
+
+            //added this due to how the ssl forwarding is working on AWS
+            if (HttpContext.Current.Request.Headers["X-Forwarded-Proto"] == "http")
+            {
+                HttpContext.Current.Response.Redirect(HttpContext.Current.Request.Url.AbsoluteUri.Replace("http:", "https:"));
+            }
+        }
+
+        /// <summary>
+        /// Ensures the SSL on normal setups.
+        /// </summary>
+        /// <param name="enableLocal">if set to <c>true</c> [enable local].</param>
+        public static void EnsureSsl(bool enableLocal = false)
+        {
+            if (HttpContext.Current.Request.Url.Host.EndsWith(".local") && !enableLocal)
+            {
+                return;
+            }
+
+            if (HttpContext.Current.Request.Url.Port != 443)
+            {
+                HttpContext.Current.Response.Redirect(HttpContext.Current.Request.Url.AbsoluteUri.Replace("http:", "https:"));
+            }
         }
     }
 }
